requirements.txt
alembic==1.7.5
appdirs==1.4.4
asgiref==3.4.1
attrs==21.2.0
autopep8==1.6.0
Babel==2.6.0
backports.entry-points-selectable==1.1.1
bcrypt==2.0.0
beautifulsoup4==4.10.0
bootstrap4==0.1.0
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
click==8.0.3
colorama==0.4.4
cryptography==36.0.0
decorator==4.3.0
defusedxml==0.7.1
distlib==0.3.3
Django==3.2.9
django-allauth==0.46.0
django-bootstrap4==21.1
django-cors-headers==3.10.0
django-cors-middleware==1.5.0
django-countries==7.2.1
django-crispy-forms==1.13.0
django-filter==21.1
django-mathfilters==1.0.0
django-phonenumber-field==6.0.0
django-rest-knox==4.1.0
django-taggit==2.0.0
django-utils-six==2.0
django-widget-tweaks==1.4.9
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.0.0
docutils==0.14
ebaysdk==2.1.5
filelock==3.4.0
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
freezegun==0.3.11
gevent==1.5.0
greenlet==0.4.15
html2text==2018.1.9
idna==2.6
importlib-metadata==2.1.2
importlib-resources==5.4.0
isodate==0.6.0
itsdangerous==2.0.1
Jinja2==3.0.3
libsass==0.17.0
lxml==4.6.4
Mako==1.0.7
MarkupSafe==2.0.1
npm==0.1.1
num2words==0.5.6
numpy==1.21.5
oauthlib==3.1.1
ofxparse==0.19
optional-django==0.1.0
pandas==1.3.5
passlib==1.7.1
phonenumbers==8.12.38
Pillow==6.1.0
platformdirs==2.4.0
polib==1.1.0
psutil==5.6.6
psycopg2==2.8.5
pycodestyle==2.8.0
pycparser==2.21
pydot==1.4.1
PyJWT==1.7.1
pyparsing==3.0.6
PyPDF2==1.26.0
pypiwin32==223
pyserial==3.4
python-dateutil==2.7.3
python-stdnum==1.14
python3-openid==3.2.0
pytz==2019.1
pyusb==1.0.2
pywin32==302
qrcode==6.1
reportlab==3.5.13
requests==2.21.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
soupsieve==2.3.1
SQLAlchemy==1.4.27
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.0.0
urllib3==1.24.3
virtualenv==20.10.0
vobject==0.9.6.1
Werkzeug==2.0.2
whitenoise==5.3.0
WTForms==3.0.0
xlrd==1.1.0
XlsxWriter==1.1.2
xlwt==1.3.0
xmltodict==0.12.0
zeep==3.2.0
zipp==3.6.0
